requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-amazon
# psycopg2-binary
# dbt-core
# dbt-postgres